Did a bit of refactoring and tidy up

+++ roguelike/roguelike/Core/Monster.cs
@@ -1,9 +1,5 @@
 ﻿using RLNET;
 using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using roguelike.Behaviors;
 using roguelike.Systems;
 
@@ -21,7 +17,7 @@ namespace roguelike.Core
             statConsole.Print(1, yPosition, Symbol.ToString(), Color);
 
             // Figure out the width of the health bar by dividing current health by max health
-            int width = Convert.ToInt32(((double)Health / (double)MaxHealth) * 16.0);
+            int width = Convert.ToInt32(((double)Health / MaxHealth) * 16.0);
             int remainingWidth = 16 - width;
 
             // Set the background colors of the health bar to show how damaged the monster is
